cache

+++ project-module-user/project-module-user-api/src/main/java/project/module/user/service/login/user/LoginUserService.java
@@ -10,6 +10,6 @@ public interface LoginUserService {
     int updateById(LoginUserPo entity);
 
 
-    LoginUserPo selectByPrimaryKey(long id);
+    LoginUserPo selectByPrimaryKey(Integer id);
 
 }
