+++ project-module-user/project-module-user-service/src/main/java/project/module/user/repository/rdb/user/LoginUserMapper.java
@@ -14,7 +14,7 @@ public interface LoginUserMapper {
 	int updateById(LoginUserPo entity);
 
 
-	LoginUserPo selectByPrimaryKey(@Param("id") long id);
+	LoginUserPo selectByPrimaryKey(@Param("id") Integer id);
