+++ project-module-user/project-module-user-service/src/main/java/project/module/user/service/login/user/impl/LoginUserServiceImpl.java
@@ -29,7 +29,7 @@ public class LoginUserServiceImpl implements LoginUserService {
     }
 
     @Override
-    public LoginUserPo selectByPrimaryKey(long id) {
+    public LoginUserPo selectByPrimaryKey(Integer id) {
         return loginUserMapper.selectByPrimaryKey(id);
     }
 }
